BackEnd/TaskFlowBackEnd/requirements.txt
Django>=5.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
